requirements.txt
alembic==1.16.4
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
gunicorn==21.2.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.9
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
reportlab==4.4.2
sendgrid==6.12.4
sib-api-v3-sdk==7.6.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
